Api cleanup part 2

+++ Api/Server/Controllers/OrderController.cs
@@ -26,36 +26,54 @@ namespace NotatnikMechanika.Server.Controllers
         [HttpGet(OrderPaths.ExtendedOrdersPath)]
         public async Task<ActionResult<Response<IEnumerable<OrderExtendedModel>>>> AllExtendedAsync(bool archived = false)
         {
+            if (User.Identity == null)
+                return Unauthorized();
+            
             return Ok(await _orderService.AllExtendedAsync(User.Identity.Name, archived));
         }
 
         [HttpGet(OrderPaths.ExtendedOrderPath)]
         public async Task<ActionResult<Response<OrderExtendedModel>>> ExtendedAsync(int orderId)
         {
+            if (User.Identity == null)
+                return Unauthorized();
+            
             return Ok(await _orderService.ExtendedAsync(User.Identity.Name, orderId, false));
         }
 
         [HttpPost(OrderPaths.AddServiceToOrderPath)]
         public async Task<ActionResult<Response>> AddServiceToOrder(int orderId, int serviceId)
         {
+            if (User.Identity == null)
+                return Unauthorized();
+            
             return Ok(await _orderService.AddServiceToOrder(User.Identity.Name, orderId, serviceId));
         }
 
         [HttpPost(OrderPaths.AddCommodityToOrderPath)]
         public async Task<ActionResult<Response>> AddCommodityToOrder(int orderId, int commodityId)
         {
+            if (User.Identity == null)
+                return Unauthorized();
+            
             return Ok(await _orderService.AddCommodityToOrder(User.Identity.Name, orderId, commodityId));
         }
 
         [HttpDelete(OrderPaths.DeleteServiceFromOrderPath)]
         public async Task<ActionResult<Response>> DeleteServiceFromOrder(int orderId, int serviceId)
         {
+            if (User.Identity == null)
+                return Unauthorized();
+            
             return Ok(await _orderService.DeleteServiceFromOrder(User.Identity.Name, orderId, serviceId));
         }
 
         [HttpDelete(OrderPaths.DeleteCommodityFromOrderPath)]
         public async Task<ActionResult<Response>> DeleteCommodityFromOrder(int orderId, int commodityId)
         {
+            if (User.Identity == null)
+                return Unauthorized();
+            
             return Ok(await _orderService.DeleteCommodityFromOrder(User.Identity.Name, orderId, commodityId));
         }
     }
